Child cell weights of LGR are now returned as double
- Automatic adding of epc extension bug
- BSpline header file not installed
