+++ CMakeLists.txt
@@ -7,7 +7,7 @@ set (FESAPI_BINARY_DIR ${CMAKE_CURRENT_BINARY_DIR})
 
 # version mechanism
 set (Fesapi_VERSION_MAJOR 0)
-set (Fesapi_VERSION_MINOR 15)
+set (Fesapi_VERSION_MINOR 16)
 set (Fesapi_VERSION_PATCH 0)
 set (Fesapi_VERSION_TWEAK 0)
 
@@ -16,7 +16,7 @@ set (Fesapi_VERSION ${Fesapi_VERSION_MAJOR}.${Fesapi_VERSION_MINOR}.${Fesapi_VER
 set (CPP_LIBRARY_NAME FesapiCpp)
 set (CS_LIBRARY_NAME FesapiCs)
 set (SUFFIX_INCLUDES_VERSION TRUE)
-set (UNDER_DEV FALSE)
+set (UNDER_DEV TRUE)
 if (${UNDER_DEV})
 	set (CPP_LIBRARY_NAME ${CPP_LIBRARY_NAME}UnderDev)
 	set (CS_LIBRARY_NAME ${CS_LIBRARY_NAME}UnderDev)
